Created generic function for fetching and processing Handlebars templates #11
