+++ views/index.html
@@ -1,17 +1,20 @@
 <!DOCTYPE html>
 <html>
-<body>
-  <h4>Welcome to the home page file</h4>
-  <h5>Your habits:</h5>
-  <ul>
-    <p>You have {{this.length}} habits</p>
-    {{#each this}}
-    <li>
-      <h5>{{title}}</h5>
-      <p>Reward: ${{reward}}</p>
-      <input type="checkbox" />
-    </li>
-    {{/each}}
-  </ul>
-</body>
+  <body>
+    <h4>Welcome to HabitBounty</h4>
+    {{#if error}}<h6>{{error}}</h6>{{/if}}
+    {{#if habits}}
+    <h5>Your habits:</h5>
+    <ul>
+      <p>You have {{this.length}} habits</p>
+      {{#each habits}}
+      <li>
+        <h5>{{title}}</h5>
+        <p>Reward: ${{reward}}</p>
+        <input type="checkbox" />
+      </li>
+      {{/each}}
+    </ul>
+    {{/if}}
+  </body>
 </html>
